refactor: consolidate duplicate users modules into canonical tree
- Remove entire src/users/ directory tree (unused users module)
- Move wallet/ and health-profile/ sub-modules to src/modules/
- Move health-profile.entity.ts to src/entities/
- Move health-profile migration to src/database/migrations/
- Update seeder and usage entity to import canonical User entity
- Update app.module.ts imports to point to new module locations
- Remove duplicate user.entity.ts (conflicted with canonical via
  TypeORM auto-load pattern)
- Remove duplicate role.enum.ts (identical to @modules/auth/enums)
- Fix missing ApiResponse import in health-profile controller

Closes #792
