+++ src/modules/health-profile/health-profile.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { HealthProfile } from '../entities/health-profile.entity';
+import { HealthProfile } from '../../entities/health-profile.entity';
 import { HealthProfileController } from './health-profile.controller';
 import { HealthProfileService } from './health-profile.service';
