+++ src/modules/usage/entities/usage.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, Column, PrimaryGeneratedColumn, CreateDateColumn, ManyToOne, JoinColumn } from 'typeorm';
-import { User } from '../../../users/entities/user.entity';
+import { User } from '../../../entities/user.entity';
 
 @Entity('usages')
 export class Usage {
