+++ src/modules/wallet/dto/wallet-summary.dto.ts
