--- src/users/enums/role.enum.ts
@@ -1,5 +1,0 @@
-export enum Role {
-  USER = 'USER',
-  HEALER = 'HEALER',
-  ADMIN = 'ADMIN',
-}
