--- src/users/users.module.ts
@@ -1,19 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
-import { User } from '../entities/user.entity';
-import { TaskCompletion } from '../tasks/entities/task-completion.entity';
-import { Coupon } from './entities/coupon.entity';
-import { WalletModule } from './wallet/wallet.module';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([User, TaskCompletion, Coupon]),
-    WalletModule,
-  ],
-  controllers: [UsersController],
-  providers: [UsersService],
-  exports: [UsersService],
-})
-export class UsersModule {}
